moved and renamed header and footer files
